CF2-004: feat: implement new components and layouts
- Added Drawer component with various subcomponents (Trigger, Portal, Close, Overlay, Content, Header, Footer, Title, Description).
- Introduced Form component with FormField, FormItem, FormLabel, FormControl, FormDescription, FormMessage, and FormField context.
- Created Input, Label, and Textarea components for form handling.
- Developed AuthLayout and DefaultLayout for routing.
- Implemented Pagination component with PaginationLink, PaginationPrevious, PaginationNext, PaginationEllipsis, and related subcomponents.
- Added Popover component with Trigger and Content.
- Created Select component with various subcomponents (Group, Value, Trigger, Content, Label, Item, Separator, Scroll buttons).
- Introduced Separator and Skeleton components for UI structure.
- Added Sonner component for notifications.
- Developed Table component with Header, Body, Footer, Row, Head, Cell, and Caption.
- Implemented Home and Login screens with respective forms and validation.
- Created PageNotFound screen for handling 404 errors.
- Added utility function for class name merging.
