+++ .prettierrc.json
@@ -1,5 +1,4 @@
 {
-  "plugins": ["@trivago/prettier-plugin-sort-imports"],
   "semi": true,
   "trailingComma": "none",
   "singleQuote": false,
