+++ src/components/Layouts/AuthLayout/AuthLayout.component.tsx
@@ -1,0 +1,10 @@
+import { Outlet } from "react-router";
+
+function AuthLayout() {
+  return (
+    <div className="bg-background min-h-screen">
+      <Outlet />
+    </div>
+  );
+}
+export { AuthLayout };
