+++ src/components/Layouts/DefaultLayout/DefaultLayout.component.tsx
@@ -1,0 +1,10 @@
+import { Outlet } from "react-router";
+
+function DefaultLayout() {
+  return (
+    <div className="bg-background min-h-screen">
+      <Outlet />
+    </div>
+  );
+}
+export { DefaultLayout };
