+++ src/components/Skeleton/Skeleton.component.tsx
@@ -1,0 +1,13 @@
+import { cn } from "@/utils/cn";
+
+function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="skeleton"
+      className={cn("bg-accent animate-pulse rounded-md", className)}
+      {...props}
+    />
+  );
+}
+
+export { Skeleton };
