+++ src/routes/ProtectedRoute/index.tsx
@@ -1,0 +1,18 @@
+import { type PropsWithChildren } from /*, useEffect*/ "react";
+
+// import { useNavigate } from "react-router-dom";
+
+type ProtectedRouteProps = PropsWithChildren;
+
+export default function ProtectedRoute({ children }: ProtectedRouteProps) {
+  //   const user = useAuth();
+  //   const navigate = useNavigate();
+
+  //   useEffect(() => {
+  //     if (user === null) {
+  //       navigate("/signin", { replace: true });
+  //     }
+  //   }, [navigate, user]);
+
+  return children;
+}
