+++ src/routes/routes.tsx
@@ -1,0 +1,8 @@
+import Home from "@/screens/Home/Home.screen";
+
+export const APP_ROUTES = [
+  {
+    path: "/",
+    element: <Home />
+  }
+];
